refactor: update links from /registry to /extensions and remove unused ExtensionCarousel component Signed-off-by: Adrian Barrio <statickidz@gmail.com>
